Houston

+++ api/Models/Manager.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace api.Models
 {
     public class Manager
@@ -5,7 +7,7 @@ namespace api.Models
         public int managerID {get; set;}
         public string mfname {get; set;}
         public string mlname {get; set;}
-        public int mphoneNumber {get; set;}
+        public Int64 mphoneNumber {get; set;}
         public string mEmail {get; set;}
         public string mUsername {get; set;}
         public string mPassword {get; set;}
